Only create a connection pool if the spec has the necessary data.
Otherwise, return the spec, as it could represent a DataSource that is
already pooled, possibly referenced by a JNDI name. We want to get the
benefits from the korma api for those kinds of connection specs as
well.

Also, we no longer assume a spec has a :pool entry, though we'll use
it if it does.