Pass driver properties to connection pool
This allows setting JDBC driver properties in the same way whether connection pool is in use or not.

There is no need to set user and password separately, they are included in the properties and are read from there.

It is not strictly necessary to dissoc irrelevant keys when setting connection pool properties.