[jdbc] Upgrade clojure.java.jdbc to 0.6.1
This commit updates clojure.java.jdbc to 0.6.1. This update is important
in that it also addresses a number of cases in which features have not
only been deprecated but officially dropped in the 0.6.x branch.

clojure.java.jdbc includes a call to `reduce-kv`, which was introduced in
Clojure 1.4.0. This means that it is not compatible with Clojure 1.3.0 or
earlier.