[fix] Fix argslist metadata for join
Although the current arglists metadata for korma.core/join is helpful to a human
user, it's technically invalid for machines. Any compiler obeying the Clojure
spec that attempts to parse the `join` macro will announce that a function
cannot have two overloads with the same arity. We might benefit from expanding
the documentation in this case to make up for the change.